src/utils/sasa_engine.py
"""
SASA Engine (Phase D: Structural Projection)
==============================================
Maps discovered modifications to 3D physical models.

Key Functions:
- SASA (Solvent Accessible Surface Area) calculation
- Residue accessibility filtering
- 3D coordinate mapping
- Sequence alignment for partial PDB structures
"""

import numpy as np
from typing import Dict, List, Tuple, Optional
from dataclasses import dataclass
from Bio import PDB
from Bio.PDB import PDBIO
import freesasa


@dataclass
class ResidueAccessibility:
    """Stores accessibility information for a residue."""
    residue_id: int  # 1-based residue number
    residue_name: str  # Three-letter code (e.g., 'LYS')
    sasa: float  # Solvent Accessible Surface Area in Å²
    is_accessible: bool  # True if SASA > threshold
    coordinates: np.ndarray  # 3D coordinates of residue


@dataclass
class StructureMapping:
    """Maps a protein sequence to a 3D structure."""
    uniprot_id: str
    pdb_id: str
    chain_id: str
    sequence_coverage: float  # Percentage of sequence covered by PDB
    residue_mappings: Dict[int, ResidueAccessibility]  # Residue number -> accessibility


class SASAEngine:
    """
    SASA Engine: Maps discovered modifications to 3D physical models.
    
    Mathematical Foundations:
    - SASA Calculation: Determines if a residue is physically reachable
    - Accessibility Filter: Residue is "Likely Target" if SASA > 20 Å²
    - Confidence Scoring: C = w₁(E_ppm) + w₂(r) + w₃(SASA)
    """
    
    SASA_THRESHOLD = 20.0  # Å² - minimum SASA for accessible residue
    MIN_COVERAGE_THRESHOLD = 0.7  # 70% sequence coverage required
    
    def __init__(self):
        """Initialize the SASA engine."""
        self.parser = PDB.PDBParser(QUIET=True)
        self.io = PDBIO()
    
    def load_pdb_structure(self, pdb_id: str, pdb_file: str = None) -> PDB.Structure:
        """
        Load a PDB structure from file or fetch from database.
        
        TODO (Task T5): Implement PDB loading
        - Load from local file if provided
        - Otherwise fetch from PDB database
        - Handle multiple models/chains
        
        Args:
            pdb_id: PDB structure ID
            pdb_file: Optional path to local PDB file
            
        Returns:
            BioPython Structure object
        """
        if pdb_file:
            structure = self.parser.get_structure(pdb_id, pdb_file)
        else:
            # TODO: Fetch from PDB database
            raise NotImplementedError("PDB fetching not yet implemented (Task T5)")
        
        return structure
    
    def calculate_sasa(self, structure: PDB.Structure, 
                      chain_id: str = None) -> Dict[int, float]:
        """
        Calculate Solvent Accessible Surface Area for each residue.
        
        Uses FreeSASA library for SASA calculation.
        
        Args:
            structure: BioPython Structure object
            chain_id: Specific chain to analyze (None for all chains)
            
        Returns:
            Dictionary mapping residue number to SASA value
        """
        # Convert BioPython structure to PDB format string
        # TODO: Implement proper conversion
        # For now, placeholder
        
        # Use FreeSASA to calculate SASA
        # sasa_result = freesasa.calc(structure)
        # Parse results and return dictionary
        
        # Placeholder implementation
        return {}
    
    def get_residue_coordinates(self, structure: PDB.Structure,
                               chain_id: str, residue_num: int) -> Optional[np.ndarray]:
        """
        Get 3D coordinates of a specific residue.
        
        Args:
            structure: BioPython Structure object
            chain_id: Chain identifier
            residue_num: Residue number (1-based)
            
        Returns:
            Array of coordinates (N, 3) for atoms in residue, or None if not found
        """
        try:
            chain = structure[0][chain_id]
            residue = chain[residue_num]
            
            coordinates = []
            for atom in residue:
                coords = atom.get_coord()
                coordinates.append(coords)
            
            return np.array(coordinates) if coordinates else None
        except (KeyError, AttributeError):
            return None
    
    def check_residue_accessibility(self, sasa: float) -> bool:
        """
        Check if a residue is accessible based on SASA threshold.
        
        A residue is considered a "Likely Target" if: SASA > 20 Å²
        
        Args:
            sasa: Solvent Accessible Surface Area in Å²
            
        Returns:
            True if residue is accessible
        """
        return sasa > self.SASA_THRESHOLD
    
    def map_sequence_to_structure(self, uniprot_sequence: str,
                                 pdb_structure: PDB.Structure,
                                 chain_id: str) -> StructureMapping:
        """
        Map UniProt sequence to PDB structure and calculate accessibility.
        
        TODO (Task T5): Implement sequence alignment
        - Align UniProt sequence with PDB sequence
        - Handle sequence mismatches and gaps
        - Calculate sequence coverage
        - Map residue numbers
        
        Args:
            uniprot_sequence: Full-length UniProt sequence
            pdb_structure: PDB structure object
            chain_id: Chain to analyze
            
        Returns:
            StructureMapping object
        """
        # Placeholder implementation
        # TODO: Implement sequence alignment using BioPython's Bio.Align
        
        return StructureMapping(
            uniprot_id="",
            pdb_id="",
            chain_id=chain_id,
            sequence_coverage=0.0,
            residue_mappings={}
        )
    
    def find_accessible_residues(self, structure: PDB.Structure,
                                chain_id: str, residue_type: str = None) -> List[ResidueAccessibility]:
        """
        Find all accessible residues of a given type in a structure.
        
        Args:
            structure: BioPython Structure object
            chain_id: Chain identifier
            residue_type: Three-letter residue code (e.g., 'LYS') or None for all
            
        Returns:
            List of ResidueAccessibility objects
        """
        accessible_residues = []
        sasa_dict = self.calculate_sasa(structure, chain_id)
        
        chain = structure[0][chain_id]
        for residue in chain:
            residue_name = residue.get_resname()
            
            # Filter by residue type if specified
            if residue_type and residue_name != residue_type:
                continue
            
            residue_num = residue.get_id()[1]
            sasa = sasa_dict.get(residue_num, 0.0)
            
            if self.check_residue_accessibility(sasa):
                coords = self.get_residue_coordinates(structure, chain_id, residue_num)
                accessible_residues.append(ResidueAccessibility(
                    residue_id=residue_num,
                    residue_name=residue_name,
                    sasa=sasa,
                    is_accessible=True,
                    coordinates=coords if coords is not None else np.array([])
                ))
        
        return accessible_residues
    
    def check_pdb_coverage(self, uniprot_sequence: str,
                          pdb_sequence: str) -> float:
        """
        Check sequence coverage of PDB structure relative to UniProt sequence.
        
        If PDB coverage < 70%, should trigger AlphaFold DB fallback.
        
        Args:
            uniprot_sequence: Full UniProt sequence
            pdb_sequence: Sequence from PDB structure
            
        Returns:
            Coverage percentage (0-1)
        """
        # Simple coverage calculation
        # TODO: Implement proper sequence alignment
        if len(uniprot_sequence) == 0:
            return 0.0
        
        # Placeholder: assume full coverage if sequences match length
        # Real implementation should use sequence alignment
        coverage = min(1.0, len(pdb_sequence) / len(uniprot_sequence))
        return coverage
    
    def get_alphafold_fallback(self, uniprot_id: str) -> Optional[str]:
        """
        Get AlphaFold structure as fallback when PDB coverage is insufficient.
        
        Triggered when PDB coverage < 70%.
        
        Args:
            uniprot_id: UniProt accession number
            
        Returns:
            Path to AlphaFold structure file or None
        """
        # TODO: Implement AlphaFold DB lookup
        # - Query AlphaFold database
        # - Download structure file
        # - Return file path
        return None

